update default agent system message
